+++ src/ffmpeg.js
@@ -56,10 +56,11 @@ function handleProgress(process, cutDuration, onProgress) {
 }
 
 async function cut({
-  customOutDir, filePath, format, cutFrom, cutTo, videoDuration, rotation, onProgress,
+  customOutDir, filePath, format, cutFrom, cutTo, videoDuration, rotation, includeAllStreams,
+  onProgress, stripAudio,
 }) {
   const ext = path.extname(filePath) || `.${format}`;
-  const cutSpecification = `${util.formatDuration(cutFrom)}-${util.formatDuration(cutTo)}`;
+  const cutSpecification = `${util.formatDuration(cutFrom, true)}-${util.formatDuration(cutTo, true)}`;
 
   const outPath = util.getOutPath(customOutDir, filePath, `${cutSpecification}${ext}`);
 
@@ -71,8 +72,12 @@ async function cut({
 
   const rotationArgs = rotation !== undefined ? ['-metadata:s:v:0', `rotate=${rotation}`] : [];
   const ffmpegArgs = [
-    '-i', filePath, '-y', '-vcodec', 'copy', '-acodec', 'copy', '-scodec', 'copy',
+    '-i', filePath, '-y',
+    ...(stripAudio ? ['-an'] : ['-acodec', 'copy']),
+    '-vcodec', 'copy',
+    '-scodec', 'copy',
     ...cutFromArgs, ...cutToArgs,
+    ...(includeAllStreams ? ['-map', '0'] : []),
     '-map_metadata', '0',
     ...rotationArgs,
     '-f', format,
@@ -90,6 +95,27 @@ async function cut({
   console.log(result.stdout);
 
   return util.transferTimestamps(filePath, outPath);
+}
+
+async function html5ify(filePath, outPath, encodeVideo) {
+  console.log('Making HTML5 friendly version', { filePath, outPath, encodeVideo });
+
+  const videoArgs = encodeVideo
+    ? ['-vf', 'scale=-2:400,format=yuv420p', '-sws_flags', 'neighbor', '-vcodec', 'libx264', '-profile:v', 'baseline', '-x264opts', 'level=3.0', '-preset:v', 'ultrafast', '-crf', '28']
+    : ['-vcodec', 'copy'];
+
+  const ffmpegArgs = [
+    '-i', filePath, ...videoArgs, '-an',
+    '-y',
+    outPath,
+  ];
+
+  console.log('ffmpeg', ffmpegArgs.join(' '));
+
+  const ffmpegPath = await getFfmpegPath();
+  const process = execa(ffmpegPath, ffmpegArgs);
+  const result = await process;
+  console.log(result.stdout);
 }
 
 /**
@@ -145,4 +171,5 @@ module.exports = {
   cut,
   getFormat,
   showFfmpegFail,
+  html5ify,
 };